requirements.txt
blinker==1.6.2
boto3==1.28.21
botocore==1.31.21
click==8.1.6
Flask==2.3.2
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
packaging==23.1
Pillow==10.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.6.1
six==1.16.0
urllib3==1.26.16
Werkzeug==2.3.6
opencv-python==4.8.0.74
torch==2.0.1
ultralytics==8.0.150